Update README.md for dependencies installation.
The installation as per the current provided dependencies is is failing with error "Do you want to continue? [Y/n] Abort.", so it isn't letting you select Y or N and its simply failing. So the correct script should have `-y` added so that the installation doesn't fail and consider Y as go ahead for installation.
I have tested it on my Ubuntu 20.4 instance and it works completely fine after adding -y flag.