Use reference vars for iteration in sysmonLogView.cpp
The prior implementation declared a loop-local copy for each of the
iterators on filters & eventIdFields in processCmdLine. Issue #23
points out that this causes a compilation failure when -Werror -Wall
is used, because this introduces an inefficiency. This change makes
these all references, which saves the superfluous copies and makes
the code compile again.

Fixes #23