Treat NetworkTracker as a class, not struct
The NetworkTracker type is declared as a class, however the code in
networkTracker.cpp attempts to refer to it as a struct. This may cause
run-time issues depending upon compiler, and also is an inconsistency
in the source code. Modify the rest of the type declarations to remove
'struct' so it is just treated like a class.