How to store/redirect Syslogs to a custom file
In this tutorial i have firstly install and configure the syslogs, to generate the system logs like file creation, file termination, process creation and network connections,
after that to redirect or store these logs to a custom file how i did the brief instruction is mention in this file must follow them and solve your problem of relating to this if you guys have. 
thanks me later.